+++ s/nexus/parts/use/parts/helpers.ts
@@ -1,12 +1,12 @@
 
-import {InitFn, InitResult, SetupFn} from "./types.js"
+import {Init, InitResult, Mount} from "./types.js"
 import {Constructor} from "../../../../tools/constructor.js"
 
 /** identity function to help you make an init function */
-export const initFn = <I extends InitFn<any>>(fn: I) => fn
+export const initFn = <I extends Init<any>>(fn: I) => fn
 
-/** identity function to help you make an setup function */
-export const setupFn = (fn: SetupFn) => fn
+/** identity function to help you make a mount function */
+export const mountFn = (fn: Mount) => fn
 
 /** a class intended to be used with use.init via initiate() function */
 export class Initiator {
@@ -29,7 +29,7 @@ export function initiator<I extends Initiator>(instance: I): InitResult<I> {
 
 /** initiate an instance of an initiator class (helpful for use.init) */
 export function initiate<C extends Constructor<Initiator>>(C: C) {
-	return (...params: ConstructorParameters<C>): InitFn<InstanceType<C>> => () => {
+	return (...params: ConstructorParameters<C>): Init<InstanceType<C>> => () => {
 		const instance = new C(...params) as InstanceType<C>
 		return [instance, () => instance.deinit()]
 	}