+++ s/nexus/parts/use/parts/utils/counter.ts
@@ -1,0 +1,13 @@
+
+export class Counter {
+	#value = 0
+
+	pull() {
+		return this.#value++
+	}
+
+	reset() {
+		this.#value = 0
+	}
+}
+